Added URl routing for tabs in buckets
