+++ portal-ui/package.json
@@ -78,7 +78,7 @@
   },
   "proxy": "http://localhost:9090/",
   "devDependencies": {
-    "prettier": "2.2.1",
+    "prettier": "2.3.0",
     "typescript": "^4.1.3"
   }
 }
